Throw exception directly
According to my (limited) understanding, without adding JGiven as a module the tests depending on it will not function. Therefore, continuing without a library is nonsensical. Furthermore if JGiven is supposed to be present it must be present as a jar (because tha's what we publish), therefore there is no use in trying maybe add Jgiven as a collection of class files (this should only be the case if we were in the same project)

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>